feat(game): 实现副本战斗系统基础功能
- 添加常量类 Contstant 用于生成随机攻击字符串- 修改 DungeonGameEntity 属性类型从 Int 为 Long- 引入日志记录和转发消息构建器- 实现基础战斗流程与回合机制- 添加伤害结算逻辑与属性影响计算- 更新 Impact 类支持伤害类型字符串和长整型数值
- 调整 Property 类属性类型为 Long 以支持更大数值范围
- 修复玩家回合分配逻辑错误- 添加战斗事件触发与效果处理框架

+++ src/main/kotlin/cn/chahuyun/omr/game/DungeonGameEntity.kt
@@ -7,9 +7,9 @@ import cn.chahuyun.omr.entity.GameEntity
  */
 class DungeonGameEntity(override var name: String = "副本", override var description: String = "副本环境") :
     GameEntity() {
-    override val hp: Int = 0
-    override val atk: Int = 0
-    override val def: Int = 0
+    override val hp: Long = 0
+    override val atk: Long = 0
+    override val def: Long = 0
     override val crit: Int = 0
     override val critDamage: Int = 0
 }
